backend/utils/token.py
from jose import jwt
from datetime import datetime, timedelta, timezone
import secrets
 
SECRET_KEY = "supersecretkey1234567890!@#$%^&*()_+-=~"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30
 
def create_access_token(data: dict, expires_delta: timedelta = None):
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + (expires_delta or timedelta(minutes=15))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)